requirements.txt
fastapi
uvicorn[standard]
databases[sqlite]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
aiosqlite